Move new TextFields into TextControls directory
