+++ MaterialComponents.podspec
@@ -1885,24 +1885,24 @@ Pod::Spec.new do |mdc|
     end
   end
 
-  mdc.subspec "TextFields+ContainedInputView" do |extension|
-    extension.ios.deployment_target = '9.0'
-    extension.public_header_files = "components/#{extension.base_name.split('+')[0]}/src/#{extension.base_name.split('+')[1]}/*.h"
-    extension.source_files = [
-    "components/#{extension.base_name.split('+')[0]}/src/#{extension.base_name.split('+')[1]}/*.{h,m}",
-    "components/#{extension.base_name.split('+')[0]}/src/#{extension.base_name.split('+')[1]}/private/*.{h,m}"
+  mdc.subspec "TextControls" do |component|
+    component.ios.deployment_target = '9.0'
+    component.public_header_files = "components/#{component.base_name}/src/*.h"
+    component.source_files = [
+      "components/#{component.base_name}/src/*.{h,m}",
+      "components/#{component.base_name}/src/private/*.{h,m}"
     ]
 
-    extension.dependency "MaterialComponents/AnimationTiming"
-    extension.dependency "MaterialComponents/private/Math"
-    extension.dependency "MDFInternationalization"
+    component.dependency "MaterialComponents/AnimationTiming"
+    component.dependency "MaterialComponents/private/Math"
+    component.dependency "MDFInternationalization"
 
-    extension.test_spec 'UnitTests' do |unit_tests|
+    component.test_spec 'UnitTests' do |unit_tests|
       unit_tests.source_files = [
-      "components/#{extension.base_name.split('+')[0]}/tests/unit/#{extension.base_name.split('+')[1]}/*.{h,m,swift}",
-      "components/#{extension.base_name.split('+')[0]}/tests/unit/#{extension.base_name.split('+')[1]}/supplemental/*.{h,m,swift}"
+        "components/#{component.base_name}/tests/unit/*.{h,m,swift}",
+        "components/#{component.base_name}/tests/unit/supplemental/*.{h,m,swift}"
       ]
-      unit_tests.resources = "components/#{extension.base_name.split('+')[0]}/tests/unit/#{extension.base_name.split('+')[1]}/resources/*"
+      unit_tests.resources = "components/#{component.base_name.split('+')[0]}/tests/unit/#{component.base_name.split('+')[1]}/resources/*"
     end
   end
