+++ catalog/Podfile
@@ -70,7 +70,7 @@ target "MDCCatalog" do
     'Tabs/UnitTests',
     'Tabs+Theming/UnitTests',
     'TextFields/UnitTests',
-    'TextFields+ContainedInputView/UnitTests',
+    'TextControls/UnitTests',
     'TextFields+Theming/UnitTests',
     'Themes/UnitTests',
     'Typography/UnitTests',
