+++ components/TextControls/examples/MDCTextControlTextFieldTypicalUseExample.m
@@ -16,7 +16,7 @@
 
 #import "MaterialButtons.h"
 #import "MaterialContainerScheme.h"
-#import "MaterialTextFields+ContainedInputView.h"
+#import "MaterialTextControls.h"
 
 static NSString *const kExampleTitle = @"MDCTextControl TextFields";
 static CGFloat const kDefaultPadding = 15.0;
