+++ components/TextControls/src/private/MDCTextControlColorViewModel.m
