+++ components/TextControls/src/private/MDCTextControlLabelAnimation.h
