+++ components/TextControls/src/private/MDCTextControlLabelAnimation.m
