+++ components/TextControls/src/private/MDCTextControlVerticalPositioningReference.h
