+++ components/TextControls/src/private/UIBezierPath+MDCTextControlStyle.h
