+++ components/TextControls/src/private/UIBezierPath+MDCTextControlStyle.m
